fix(memory): break backfill lifecycle import cycle
