+++ extensions/memory-core/src/session-backfill-contract.ts
@@ -1,0 +1,40 @@
+export type SessionBackfillDay = {
+  day: string;
+  candidateCount: number;
+  topCandidates: string[];
+};
+
+export type SessionBackfillBatchProgress = {
+  batch: number;
+  days: number;
+  candidates: number;
+  stagedEntries: number;
+};
+
+export type SessionBackfillResult = {
+  agentId: string;
+  workspaceDir: string;
+  applied: boolean;
+  rem: boolean;
+  days: SessionBackfillDay[];
+  candidateCount: number;
+  stagedEntries: number;
+  writtenDiaryEntries: number;
+  replacedDiaryEntries: number;
+  batchCount?: number;
+  batches?: SessionBackfillBatchProgress[];
+  rollback?: {
+    removedDiaryEntries: number;
+    removedStagedEntries: number;
+  };
+};
+
+export type SessionBackfillContinuation = {
+  advanced: boolean;
+  hasMore: boolean;
+};
+
+export type SessionBackfillExecution = {
+  result: SessionBackfillResult;
+  continuation: SessionBackfillContinuation;
+};
